+++ Frends.AzureDataLake.DownloadFiles/Frends.AzureDataLake.DownloadFiles/Definitions/Input.cs
@@ -30,7 +30,7 @@ public class Input
     /// </summary>
     /// <example>c:\temp</example>
     [DisplayFormat(DataFormatString = "Text")]
-    public string Directory { get; init; }
+    public string DestinationDirectory { get; init; }
 
     /// <summary>
     /// How the existing file will be handled.